separate restaurant to dish and cook app
