+++ cook/forms.py
@@ -1,0 +1,30 @@
+from django.contrib.auth.forms import UserCreationForm
+from django import forms
+from .models import Cook
+
+
+class CookCreateForm(UserCreationForm):
+    class Meta:
+        model = Cook
+        fields = UserCreationForm.Meta.fields + (
+            "first_name",
+            "last_name",
+            "years_of_experience",
+        )
+
+
+class CookUpdateForm(forms.ModelForm):
+    class Meta:
+        model = Cook
+        fields = ("first_name", "last_name", "years_of_experience",)
+
+
+class CookSearchForm(forms.Form):
+    username = forms.CharField(
+        max_length=255,
+        required=False,
+        label="",
+        widget=forms.TextInput(
+            attrs={"placeholder": "Search by username"}
+        )
+    )
