+++ dish/admin.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+from .models import DishType, Dish
+
+
+@admin.register(DishType)
+class DishTypeAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(Dish)
+class DishAdmin(admin.ModelAdmin):
+    list_display = ["name", "description", "price", "dish_type"]
+    list_filter = ["dish_type__name", ]
+    search_fields = ["name", ]
