+++ dish/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class RestaurantConfig(AppConfig):
+class DishConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "restaurant"
+    name = "dish"
