+++ dish/forms.py
@@ -1,23 +1,6 @@
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm
 from django import forms
-from restaurant.models import Cook, Dish
-
-
-class CookCreateForm(UserCreationForm):
-    class Meta:
-        model = Cook
-        fields = UserCreationForm.Meta.fields + (
-            "first_name",
-            "last_name",
-            "years_of_experience",
-        )
-
-
-class CookUpdateForm(forms.ModelForm):
-    class Meta:
-        model = Cook
-        fields = ("first_name", "last_name", "years_of_experience",)
+from .models import Dish
 
 
 class DishForm(forms.ModelForm):
@@ -38,17 +21,6 @@ class DishSearchForm(forms.Form):
         label="",
         widget=forms.TextInput(
             attrs={"placeholder": "Search by name"}
-        )
-    )
-
-
-class CookSearchForm(forms.Form):
-    username = forms.CharField(
-        max_length=255,
-        required=False,
-        label="",
-        widget=forms.TextInput(
-            attrs={"placeholder": "Search by username"}
         )
     )
