+++ kitchen_service/urls.py
@@ -21,6 +21,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("", include("restaurant.urls", namespace="restaurant")),
+    path("", include("dish.urls", namespace="dish")),
+    path("", include("cook.urls", namespace="cook")),
     path("accounts/", include("django.contrib.auth.urls")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
