+++ templates/cook/cook_list.html
@@ -27,7 +27,7 @@
                     <h4 class="mb-0">{{ cook.username}}</h4>
                     <div class="d-block">
                       <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                        <a href="{% url 'restaurant:cook-detail' pk=cook.id %}">Detail Information</a>
+                        <a href="{% url 'cook:cook-detail' pk=cook.id %}">Detail Information</a>
                       </button>
                     </div>
                   </div>
