+++ templates/dish/dish_detail.html
@@ -21,7 +21,7 @@
             <div class="col-lg-7 col-md-7 z-index-2 position-relative px-md-2 px-sm-5 mt-sm-0 mt-4">
               <div class="d-flex justify-content-between align-items-center mb-2">
                 <h4 class="mb-0">{{ dish.name}}</h4>
-                <form method="post" action="{% url 'restaurant:dish-cook-update' pk=dish.id %}">
+                <form method="post" action="{% url 'dish:dish-cook-update' pk=dish.id %}">
                   {% csrf_token %}
                   {% if user in dish.cooks.all %}
                     <div class="d-block">
@@ -56,7 +56,7 @@
                 <hr>
                 <div class="d-flex justify-content-between align-items-center mb-2">
                     <h4 class="mb-0">
-                      <a href="{% url 'restaurant:cook-detail' pk=cook.id %}">{{ cook.username}}</a></h4>
+                      <a href="{% url 'cook:cook-detail' pk=cook.id %}">{{ cook.username}}</a></h4>
                 </div>
                 <div class="row mb-4">
                   <div class="col-auto">
@@ -72,10 +72,10 @@
               {% endfor %}
                 <div class="d-block">
                   <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                    <a href="{% url 'restaurant:dish-update' pk=dish.id %}">Update Dish</a>
+                    <a href="{% url 'dish:dish-update' pk=dish.id %}">Update Dish</a>
                   </button>
                   <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                    <a href="{% url 'restaurant:dish-delete' pk=dish.id %}">Delete Dish</a>
+                    <a href="{% url 'dish:dish-delete' pk=dish.id %}">Delete Dish</a>
                   </button>
                 </div>
             </div>
