+++ templates/dish/dish_list.html
@@ -26,7 +26,7 @@
                     <h4 class="mb-0">{{ dish.name}}</h4>
                     <div class="d-block">
                       <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                        <a href="{% url 'restaurant:dish-detail' pk=dish.id %}">Detail Information</a>
+                        <a href="{% url 'dish:dish-detail' pk=dish.id %}">Detail Information</a>
                       </button>
                     </div>
                   </div>
