+++ templates/dish/dishtype_list.html
@@ -26,10 +26,10 @@
                     <h4 class="mb-0">{{ dishtype.name}}</h4>
                     <div class="d-block">
                       <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                        <a href="{% url 'restaurant:dish_type-update' pk=dishtype.id %}">Update</a>
+                        <a href="{% url 'dish:dish_type-update' pk=dishtype.id %}">Update</a>
                       </button>
                       <button type="button" class="btn btn-sm btn-outline-info text-nowrap mb-0">
-                        <a href="{% url 'restaurant:dish_type-delete' pk=dishtype.id %}">Delete</a>
+                        <a href="{% url 'dish:dish_type-delete' pk=dishtype.id %}">Delete</a>
                       </button>
                     </div>
                   </div>
