+++ templates/includes/navigation.html
@@ -32,10 +32,10 @@
                   aria-labelledby="dropdownMenuPages">
                   <div class="d-none d-lg-block">
                     {% if request.user.is_authenticated %}
-                    <a href="{% url 'restaurant:dish-list' %}" class="dropdown-item border-radius-md">
+                    <a href="{% url 'dish:dish-list' %}" class="dropdown-item border-radius-md">
                       <span class="ps-3">Dishes</span>
                     </a>
-                    <a href="{% url 'restaurant:dish-create' %}" class="dropdown-item border-radius-md">
+                    <a href="{% url 'dish:dish-create' %}" class="dropdown-item border-radius-md">
                       <span class="ps-3">Add new dish</span>
                     </a>
                     {% else %}
@@ -63,7 +63,7 @@
                   </div>
                   <div class="d-none d-lg-block">
                     <li class="nav-item dropdown dropdown-hover dropdown-subitem">
-                      <a href="{% url 'restaurant:dish_type-list' %}" class="dropdown-item border-radius-md">
+                      <a href="{% url 'dish:dish_type-list' %}" class="dropdown-item border-radius-md">
                         <span class="ps-3">Dish Types</span>
                       </a>
                     </li>
@@ -71,7 +71,7 @@
 
 
                     <li class="nav-item dropdown dropdown-hover dropdown-subitem">
-                      <a href="{% url 'restaurant:dish_type-create' %}" class="dropdown-item border-radius-md">
+                      <a href="{% url 'dish:dish_type-create' %}" class="dropdown-item border-radius-md">
                         <span class="ps-3">Add new Dish Type</span>
                       </a>
                     </li>
@@ -90,18 +90,18 @@
                   <div class="d-none d-lg-block">
                     <ul class="list-group">
                       <li class="nav-item list-group-item border-0 p-0">
-                        <a href="{% url 'restaurant:cook-list' %}" class="dropdown-item border-radius-md">
+                        <a href="{% url 'cook:cook-list' %}" class="dropdown-item border-radius-md">
                           <span class="ps-3">Cooks</span>
                         </a>
                       </li>
                       <li class="nav-item list-group-item border-0 p-0">
-                        <a href="{% url 'restaurant:cook-create' %}" class="dropdown-item border-radius-md">
+                        <a href="{% url 'cook:cook-create' %}" class="dropdown-item border-radius-md">
                           <span class="ps-3">Add new Cook</span>
                         </a>
                       </li>
                       {% if user.is_authenticated %}
                       <li class="nav-item list-group-item border-0 p-0">
-                        <a href="{% url 'restaurant:cook-update' pk=user.id %}" class="dropdown-item border-radius-md">
+                        <a href="{% url 'cook:cook-update' pk=user.id %}" class="dropdown-item border-radius-md">
                           <span class="ps-3">Change my Data</span>
                         </a>
                       </li>
@@ -109,7 +109,7 @@
                       <li class="nav-item list-group-item border-0 p-0">
                         {% if user.is_authenticated %}
                         <li class="nav-item list-group-item border-0 p-0">
-                          <a href="{% url 'restaurant:cook-password-update' pk=user.id %}" class="dropdown-item border-radius-md">
+                          <a href="{% url 'cook:cook-password-update' pk=user.id %}" class="dropdown-item border-radius-md">
                             <span class="ps-3">Change my Password</span>
                           </a>
                       </li>
