+++ templates/registration/login.html
@@ -46,7 +46,7 @@
             <div class="card-footer text-center pt-0 px-lg-2 px-1">
               <p class="mb-4 text-sm mx-auto">
                 Don't have an account?
-                <a href="{% url 'restaurant:cook-create' %}" class="text-primary text-gradient font-weight-bold">Sign up</a>
+                <a href="{% url 'cook:cook-create' %}" class="text-primary text-gradient font-weight-bold">Sign up</a>
               </p>
             </div>
           </div>
